Added interface to allow different printing styles.
This change is a no-op for existing goblin users.  This is only
useful if you are running in a terminal that doesn't understand
the ansi color codes. This matters to users of go sublime because
sublime isnt a terminal so the colors and check make the output
hard to read.