Accept various types of messages in assertions
Expand all assertions to accept various types of messages to emit as
output when the expression evaluates to false.  This changes permits
unit tests to use any assertion call for returning a message instead of
just IsTrue() and IsFalse().

The change from accepting only strings to any value allows for
simplification of unit tests that just print the error string.

This allows replacing:
if err != nil {
	g.Fail(err.Error())
}

With:
g.Assert(err != nil).IsTrue(err)

The logic to format the messages into output relies upon fmt.Sprintf().
It simply concatenates the messages together with one exception:
messages that are only whitespace.  In this case, it wraps the
whitespace with square brackets to make it more obvious that a message
was provided albeit unseen.