Skip.Describe support
Add the ability to skip all tests within a describe block if it was added using g.Skip.Describe().
The before and after hooks on a skipped test in this Skip.Describe block behave the same way as g.Xit() tests - they do not run. Each Describe and It block within a Skip.Describe() will be logged in yellow.